Defer numba guvectorize compilation for aggregate functions
