+++ sgkit/stats/aggregation.py
@@ -1,4 +1,3 @@
-import math
 from typing import Any, Dict, Hashable
 
 import dask.array as da
@@ -8,7 +7,6 @@ from typing_extensions import Literal
 from xarray import Dataset
 
 from sgkit import variables
-from sgkit.accelerate import numba_guvectorize, numba_jit
 from sgkit.display import genotype_as_bytes
 from sgkit.stats.utils import cohort_sum
 from sgkit.typing import ArrayLike
@@ -20,45 +18,6 @@ from sgkit.utils import (
 )
 
 Dimension = Literal["samples", "variants"]
-
-
-@numba_guvectorize(  # type: ignore
-    [
-        "void(int8[:], uint8[:], uint8[:])",
-        "void(int16[:], uint8[:], uint8[:])",
-        "void(int32[:], uint8[:], uint8[:])",
-        "void(int64[:], uint8[:], uint8[:])",
-    ],
-    "(k),(n)->(n)",
-)
-def count_alleles(
-    g: ArrayLike, _: ArrayLike, out: ArrayLike
-) -> None:  # pragma: no cover
-    """Generalized U-function for computing per sample allele counts.
-
-    Parameters
-    ----------
-    g
-        Genotype call of shape (ploidy,) containing alleles encoded as
-        type `int` with values < 0 indicating a missing allele.
-    _
-        Dummy variable of type `uint8` and shape (alleles,) used to
-        define the number of unique alleles to be counted in the
-        return value.
-
-    Returns
-    -------
-    ac : ndarray
-        Allele counts with shape (alleles,) and values corresponding to
-        the number of non-missing occurrences of each allele.
-
-    """
-    out[:] = 0
-    n_allele = len(g)
-    for i in range(n_allele):
-        a = g[i]
-        if a >= 0:
-            out[a] += 1
 
 
 def count_call_alleles(
@@ -115,6 +74,8 @@ def count_call_alleles(
            [[2, 0],
             [2, 0]]], dtype=uint8)
     """
+    from .aggregation_numba_fns import count_alleles
+
     variables.validate(ds, {call_genotype: variables.call_genotype_spec})
     n_alleles = ds.dims["alleles"]
     G = da.asarray(ds[call_genotype])
@@ -288,104 +249,6 @@ def call_rate(ds: Dataset, dim: Dimension, call_genotype_mask: Hashable) -> Data
     )
 
 
-@numba_jit(nogil=True)
-def _biallelic_genotype_index(genotype: ArrayLike) -> int:
-    index = 0
-    for i in range(len(genotype)):
-        a = genotype[i]
-        if a < 0:
-            if a < -1:
-                raise ValueError("Mixed-ploidy genotype indicated by allele < -1")
-            return -1
-        if a > 1:
-            raise ValueError("Allele value > 1")
-        index += a
-    return index
-
-
-@numba_guvectorize(  # type: ignore
-    [
-        "void(int8[:,:], uint64[:], uint64[:])",
-        "void(int16[:,:], uint64[:], uint64[:])",
-        "void(int32[:,:], uint64[:], uint64[:])",
-        "void(int64[:,:], uint64[:], uint64[:])",
-    ],
-    "(n, k),(g)->(g)",
-)
-def _count_biallelic_genotypes(
-    genotypes: ArrayLike, _: ArrayLike, out: ArrayLike
-) -> ArrayLike:  # pragma: no cover
-    out[:] = 0
-    for i in range(len(genotypes)):
-        index = _biallelic_genotype_index(genotypes[i])
-        if index >= 0:
-            out[index] += 1
-
-
-# implementation from github.com/PlantandFoodResearch/MCHap
-# TODO: replace with math.comb when supported by numba
-@numba_jit(nogil=True)
-def _comb(n: int, k: int) -> int:
-    if k > n:
-        return 0
-    r = 1
-    for d in range(1, k + 1):
-        gcd_ = math.gcd(r, d)
-        r //= gcd_
-        r *= n
-        r //= d // gcd_
-        n -= 1
-    return r
-
-
-_COMB_REP_LOOKUP = np.array(
-    [[math.comb(max(0, n + k - 1), k) for k in range(11)] for n in range(11)]
-)
-_COMB_REP_LOOKUP[0, 0] = 0  # special case
-
-
-@numba_jit(nogil=True)
-def _comb_with_replacement(n: int, k: int) -> int:
-    if (n < _COMB_REP_LOOKUP.shape[0]) and (k < _COMB_REP_LOOKUP.shape[1]):
-        return _COMB_REP_LOOKUP[n, k]
-    n = n + k - 1
-    return _comb(n, k)
-
-
-@numba_jit(nogil=True)
-def _sorted_genotype_index(genotype: ArrayLike) -> int:
-    # Warning: genotype alleles must be sorted in ascending order!
-    if genotype[0] < 0:
-        if genotype[0] < -1:
-            raise ValueError("Mixed-ploidy genotype indicated by allele < -1")
-        return -1
-    index = 0
-    for i in range(len(genotype)):
-        a = genotype[i]
-        index += _comb_with_replacement(a, i + 1)
-    return index
-
-
-@numba_guvectorize(  # type: ignore
-    [
-        "void(int8[:,:], uint64[:], uint64[:])",
-        "void(int16[:,:], uint64[:], uint64[:])",
-        "void(int32[:,:], uint64[:], uint64[:])",
-        "void(int64[:,:], uint64[:], uint64[:])",
-    ],
-    "(n, k),(g)->(g)",
-)
-def _count_sorted_genotypes(
-    genotypes: ArrayLike, _: ArrayLike, out: ArrayLike
-) -> ArrayLike:  # pragma: no cover
-    # Warning: genotype alleles must be sorted in ascending order!
-    out[:] = 0
-    for i in range(len(genotypes)):
-        index = _sorted_genotype_index(genotypes[i])
-        if index >= 0:
-            out[index] += 1
-
-
 def count_variant_genotypes(
     ds: Dataset,
     *,
@@ -457,6 +320,12 @@ def count_variant_genotypes(
            [0, 2, 0],
            [2, 0, 0]], dtype=uint64)
     """
+    from .aggregation_numba_fns import (
+        _comb_with_replacement,
+        _count_biallelic_genotypes,
+        _count_sorted_genotypes,
+    )
+
     ds = define_variable_if_absent(
         ds,
         variables.genotype_id,
@@ -509,52 +378,34 @@ def count_variant_genotypes(
     return conditional_merge_datasets(ds, new_ds, merge)
 
 
-@numba_guvectorize(  # type: ignore
-    [
-        "void(int64[:], int8[:], int8[:])",
-        "void(int64[:], int16[:], int16[:])",
-        "void(int64[:], int32[:], int32[:])",
-        "void(int64[:], int64[:], int64[:])",
-    ],
-    "(),(k)->(k)",
-)
-def _index_as_genotype(
-    index: int, _: ArrayLike, out: ArrayLike
-) -> None:  # pragma: no cover
-    """Convert the integer index of a genotype to a
-    genotype call following the VCF specification
-    for fields of length G.
+def _genotype_as_bytes(genotype: ArrayLike, max_allele_chars: int = 2) -> ArrayLike:
+    """Convert integer encoded genotype calls to (unphased)
+    VCF style byte strings.
 
     Parameters
     ----------
-    index
-        Index of genotype following the sort order described in the
-        VCF spec. An index less than 0 is invalid and will return an
-        uncalled genotype.
-    _
-        Dummy variable of length ploidy. The dtype of this variable is
-        used as the dtype of the returned genotype array.
+    genotype
+        Genotype call.
+    max_allele_chars
+        Maximum number of chars required for any allele.
+        This should include signed sentinel values.
 
     Returns
     -------
-    genotype
-        Integer alleles of the genotype call.
+    genotype_string
+        Genotype encoded as byte string.
     """
-    ploidy = len(out)
-    remainder = index
-    for index in range(ploidy):
-        # find allele n for position k
-        p = ploidy - index
-        n = -1
-        new = 0
-        prev = 0
-        while new <= remainder:
-            n += 1
-            prev = new
-            new = _comb_with_replacement(n, p)
-        n -= 1
-        remainder -= prev
-        out[p - 1] = n
+    from .aggregation_numba_fns import _format_genotype_bytes
+
+    ploidy = genotype.shape[-1]
+    b = genotype.astype("|S{}".format(max_allele_chars))
+    b.dtype = np.uint8
+    n_num = b.shape[-1]
+    n_char = n_num + ploidy - 1
+    dummy = np.empty(n_char, np.uint8)
+    c = _format_genotype_bytes(b, ploidy, dummy)
+    c.dtype = "|S{}".format(n_char)
+    return np.squeeze(c)
 
 
 def genotype_coords(
@@ -588,6 +439,8 @@ def genotype_coords(
     A dataset containing :data:`sgkit.variables.genotype_id_spec`
     containing all possible genotype strings.
     """
+    from .aggregation_numba_fns import _comb_with_replacement, _index_as_genotype
+
     n_alleles = ds.dims["alleles"]
     ploidy = ds.dims["ploidy"]
     n_genotypes = _comb_with_replacement(n_alleles, ploidy)
